Add MC statistics cheat option.
We cannot expect to have a MC simulation
that perfectly describes nature.
It is rather simple to show the effect
from the limited number of MC statistics.
For this, we simply use a seperate MC sevent set for getting anything
that should in the end be obtained from experimental data.

This is in general the right thing to do,
and should stay the default.
There are cases though in which we would
explicitly like to _not_ show the effects
that stem from limited MC statistics.
This is especially true now, where we are only using
a small fraction of the available MC statistics
from ILD central simulations.